+++ packages/build-common/src/packageMetadata/v1.ts
@@ -30,14 +30,39 @@
  * - Backwards: The reader for metadata version 1.x.\_ must be able to read metadata version 1.y.\_ if x >= y.
  * - Forwards: The reader for metadata version 1.x.y must also be able to read metadata version 1.x.z if z >= y.
  *
+ * ## Version history
+ *
+ * ### 1.1.0
+ *
+ * - Packages may now use the `open-pioneer:deployment` module.
+ *   This import is passed through during package compilation and must be handled by the vite plugin at runtime.
+ *
+ * ### 1.0.1
+ *
+ * - New optional `runtimeMeta` field.
+ *   This is used by `@open-pioneer/runtime` to indicate which version of app metadata it supports.
+ *   It should not be used by other packages.
+ *
+ * ### 1.0.0
+ *
+ * Initial release
  *
  * @module
  */
+import { SemVer } from "semver";
 import type { PackageMetadataV1 as V1 } from "../../types";
 import { canParse } from "../versionUtils";
 import { z } from "zod";
 
-export const CURRENT_VERSION = "1.1.0";
+export const LATEST_VERSION = "1.1.0";
+
+// Target (minor version) to semver with patch version (if any).
+const LATEST_VERSION_FOR_TARGET: Record<V1.MinorVersion, string> = {
+    ["1.0" as V1.MinorVersion]: "1.0.1",
+    ["1.1" as V1.MinorVersion]: "1.1.0"
+};
+
+export const MINOR_VERSIONS = ["1.0" as V1.MinorVersion, "1.1" as V1.MinorVersion] as const;
 
 /* NOTE: do not use .strict() for objects here to allow future additions of optional properties */
 
@@ -49,7 +74,7 @@ const VERSION_SCHEMA = z.object({
 
 const PROPERTY_CONFIG_SCHEMA: z.ZodType<V1.PropertyConfig> = z.object({
     propertyName: z.string(),
-    value: z.any().nullish().optional(),
+    defaultValue: z.any().nullish().optional(),
     required: z.boolean().nullish().optional()
 });
 
@@ -93,6 +118,39 @@ const PACKAGE_METADATA_SCHEMA: z.ZodType<V1.PackageMetadata> = VERSION_SCHEMA.ex
     runtimeMeta: RUNTIME_META_CONFIG_SCHEMA.nullish().optional()
 });
 
+const featuresSince: Record<"app-deployment-module", V1.MinorVersion> = {
+    "app-deployment-module": "1.1" as V1.MinorVersion
+};
+
+export const supportsFeature: typeof V1.supportsFeature = (compileTarget, feature) => {
+    const featureTarget = featuresSince[feature];
+    if (!featureTarget) {
+        throw new Error(`Unknown feature: '${feature}'`);
+    }
+
+    const featureSemver = toSemver(featureTarget);
+    const compileSemver = toSemver(compileTarget);
+    if (compileSemver.compare(featureSemver) >= 0) {
+        return {
+            supports: true
+        };
+    } else {
+        return {
+            supports: false,
+            needed: featureTarget
+        };
+    }
+};
+
+const MINOR_VERSION_RE = /^\d+\.\d+$/;
+
+function toSemver(target: V1.MinorVersion): SemVer {
+    if (!target.match(MINOR_VERSION_RE)) {
+        throw new Error(`Target '${target}' is not a valid target version.`);
+    }
+    return new SemVer(target + ".0");
+}
+
 export const parsePackageMetadata: typeof V1.parsePackageMetadata = (jsonValue) => {
     // Require that at least the version field is present.
     const versionResult = VERSION_SCHEMA.safeParse(jsonValue);
@@ -108,7 +166,7 @@ export const parsePackageMetadata: typeof V1.parsePackageMetadata = (jsonValue) 
     // Check whether the version is supported.
     const serializedVersion = versionResult.data[VERSION_FIELD];
     try {
-        if (!canParse(CURRENT_VERSION, serializedVersion)) {
+        if (!canParse(LATEST_VERSION, serializedVersion)) {
             return {
                 type: "error",
                 code: "unsupported-version",
@@ -143,18 +201,22 @@ export const parsePackageMetadata: typeof V1.parsePackageMetadata = (jsonValue) 
 };
 
 export const serializePackageMetadata: typeof V1.serializePackageMetadata = (
-    metadata: V1.OutputPackageMetadata & Partial<Pick<V1.PackageMetadata, typeof VERSION_FIELD>>
+    metadata: V1.OutputPackageMetadata,
+    target: V1.MinorVersion
 ) => {
-    if (metadata[VERSION_FIELD] != null && metadata[VERSION_FIELD] !== CURRENT_VERSION) {
-        throw new Error(
-            `Invalid package metadata version '${metadata[VERSION_FIELD]}': ` +
-                `version should either be omitted or be equal to the current version.`
-        );
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    if ((metadata as any)[VERSION_FIELD]) {
+        throw new Error(`The package metadata version should not be specified directly.`);
+    }
+
+    const latestVersion = LATEST_VERSION_FOR_TARGET[target];
+    if (!latestVersion) {
+        throw new Error(`Unknown target version: '${target}'.`);
     }
 
     const augmentedMetadata: V1.PackageMetadata = {
         ...metadata,
-        [VERSION_FIELD]: CURRENT_VERSION
+        [VERSION_FIELD]: latestVersion
     };
 
     // Sanity check: pass our own validation.