+++ packages/build-package/src/generatePackageJson.test.ts
@@ -1,11 +1,11 @@
 // SPDX-FileCopyrightText: 2023-2025 Open Pioneer project (https://github.com/open-pioneer)
 // SPDX-License-Identifier: Apache-2.0
+import { BuildConfig, DEFAULT_PACKAGE_TARGET } from "@open-pioneer/build-common";
 import { describe, expect, it } from "vitest";
 import { GeneratePackageJsonOptions, generatePackageJson } from "./generatePackageJson";
-import { createMemoryLogger } from "./utils/Logger";
 import { createInputModelFromData } from "./model/InputModel";
 import { NormalizedEntryPoint } from "./utils/entryPoints";
-import { BuildConfig } from "@open-pioneer/build-common";
+import { createMemoryLogger } from "./utils/Logger";
 import { ValidationReporter } from "./utils/ValidationReporter";
 
 describe("generatePackageJson", function () {
@@ -33,7 +33,7 @@ describe("generatePackageJson", function () {
               "i18n": {
                 "languages": [],
               },
-              "packageFormatVersion": "1.1.0",
+              "packageFormatVersion": "1.0.1",
               "properties": [],
               "services": [],
               "ui": {
@@ -273,7 +273,7 @@ describe("generatePackageJson", function () {
                 "de",
               ],
             },
-            "packageFormatVersion": "1.1.0",
+            "packageFormatVersion": "1.0.1",
             "properties": [
               {
                 "defaultValue": 1,
@@ -346,7 +346,7 @@ describe("generatePackageJson", function () {
               "i18n": {
                 "languages": [],
               },
-              "packageFormatVersion": "1.1.0",
+              "packageFormatVersion": "1.0.1",
               "properties": [],
               "runtimeMeta": {
                 "metadataVersion": "1.1.123456",
@@ -400,6 +400,7 @@ function testDefaults(options?: {
             requireReadme: true
         },
         logger,
+        packageFormatTarget: DEFAULT_PACKAGE_TARGET,
         reporter: new ValidationReporter(logger, options?.strict ?? true)
     } satisfies GeneratePackageJsonOptions;
 }