+++ packages/vite-plugin/src/metadata/MetadataRepository.ts
@@ -342,7 +342,7 @@ function detectRuntimeMetadataVersion(
     }
 
     const versionResult = RuntimeSupport.getSupportedRuntimeMetadataVersion(metadataVersion);
-    if (typeof versionResult != "string" &&  "code" in versionResult) {
+    if (typeof versionResult != "string" && "code" in versionResult) {
         switch (versionResult.code) {
             case "invalid-version":
                 throw new ReportableError(`Invalid metadata version ${metadataVersion}`, {
@@ -355,11 +355,7 @@ function detectRuntimeMetadataVersion(
                 );
         }
     }
-    isDebug &&
-    debug(
-        "Detected runtime version %s",
-        versionResult
-    );
+    isDebug && debug("Detected runtime version %s", versionResult);
     return versionResult as RuntimeSupport.RuntimeMetadataVersion;
 }