+++ packages/vite-plugin/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog @open-pioneer/vite-plugin-pioneer
 
+## 5.1.2
+
+### Patch Changes
+
+- 50ddcbd: Fix default package property values not being applied correctly when using a trails package from node modules.
+- Updated dependencies [50ddcbd]
+    - @open-pioneer/build-common@3.2.2
+
+## 5.1.1
+
+### Patch Changes
+
+- fa84b0e: Bump dependencies
+- 093ae42: Prevent detection of duplicated packages on windows
+- Updated dependencies [fa84b0e]
+    - @open-pioneer/build-common@3.2.1
+
 ## 5.1.0
 
 ### Minor Changes